+++ src/constants.rs
@@ -1,0 +1,3 @@
+pub const QUALIFIER: &'static str = "dev";
+pub const ORGANIZATION: &'static str = "CharlieKarafotias";
+pub const APPLICATION: &'static str = "Astra";